Limited agent's actions to servers
The agents' actions are indexes of the list of servers. Then there is a dictionary keeping the mapping between index and node id.
e.g. Nodes:
  - ID 0: server
  - ID 1: router
  - ID 5: router
  - ID 7: server
servers_list = [ID_0, ID_7]
map(index, id) = {0: ID_0, 1: ID_7}

THe agent will output either 0 or 1, but they correspond to the physical nodes 0 and 7.
There is an apposite dictionary keeping this map