Removed 'limited' attribute from the agents
The differentiation between limited and non-limited (i.e., can choose even nodes with not enough available resources and make the NSPR be rejected) agents has to be done in the environment.
This because a limited agent can decide not to pick any node (if there is no node with enough resources), this is equivalent to a specific action and actions are defined in the env.
So a non-limited agent will have a wrapped environment where the action space doesn't have this "no pick" action