If no physical path satisfies BW reqs -> unsuccessful action
This because networkx.shortest_path, if not link has enough BW, then all of them have infinity as weight, so all are equal and the function will return a path anyway, but it's an invalid one.
So I added a check after choosing a physical path.

Plus I fixed a bug in reader.py:
on read_nsprs, even when passing the path of a single file, the function was extracting the directory path and reading ALL the files in it, not only the one specified by the argument