Implemented PSN reset with load (first version)
The load can either be a float or a dict.
If it's a float, all the resources are initialized with that amount of load (it represents a percentage);
if it's a dict, it indicates the init load percentage of each resource independently.

Currently it's a fixed load, every time the PSN gets reset with the same load percentage.

Next I plan to implement a random amount of load sampled from a distribution whose parameters are passed to the wrapper