Fixed bug in simulator that was producing CPU availabilities > than the capacities
This was due to the fact that, when removing NSPRs from the PSN, the resources were sometimes restored twice for the same node.
The origin of this was the fact that the time step in the simulator was not increased when the actions weren't successful,and this led to synchronization problems between the removal of NSPRs because of not enough available resources and when the departure time had come. The departure time of course wasn't reached correctly, since the time step itself was wrong