NSPRs' DepartureTime now SET ONLY when the NSPR gets picked by the simulator
Previously the departure time was set a priori depending on the request's arrival time and its duration, but with high arrival rates the NSPRs' queue was getting long and request were departing before being evaluated.
Now the departure time gets set as: NSPR's picking time + its duration.

This is also good because - since the arrival rate calculation doesn't take into account that the agent takes a certain amount of steps to place the request - the arrival rate is an over-estimate. If requests depart super early (because they had to wait in the queue for long), the effective load of the PSN is even lower. Therefore, this modification brings us nearer to the target PSN load